Slithgly enhance 'ButtonPlayground'. Refactor 'Buttons' into 'Button' overall
