+++ playgrounds/TagPlayground.js
@@ -40,6 +40,7 @@ export default class TagPlayground extends React.Component {
         DropdownTag,
       },
       currentComponent: "SimpleTag",
+      ...this.initialState,
       selectedText: this.initialState.selectedText,
       selectedColor: this.initialState.selectedColor,
       selectedIsDisabled: this.initialState.selectedIsDisabled,
@@ -127,7 +128,7 @@ export default class TagPlayground extends React.Component {
         : this.handleToggleTag;
 
     return (
-      <>
+      <React.Fragment>
         {removableTagIsOn ? (
           <Component
             text={selectedText}
@@ -143,7 +144,7 @@ export default class TagPlayground extends React.Component {
         ) : (
           ""
         )}
-      </>
+      </React.Fragment>
     );
   };
