+++ stories/Buttons.stories.mdx
@@ -1,7 +1,7 @@
 import { Fragment } from 'react';
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs/blocks';
 
-import { Button, OutlineButton, TextButton } from '../labsystem/src/Buttons';
+import { Button, OutlineButton, TextButton } from '../labsystem/src/Button';
 
 import ButtonPlayground from "../playgrounds/ButtonPlayground";
 
@@ -9,7 +9,7 @@ import ButtonPlayground from "../playgrounds/ButtonPlayground";
 
 <div class="docs-title">
   <div>
-    <h1 class="docs-title__text">Buttons</h1>
+    <h1 class="docs-title__text">Button</h1>
   </div>
 </div>
