+++ stories/Tooltip.stories.mdx
@@ -2,7 +2,7 @@ import { Fragment } from "react";
 import { Meta, Story, Canvas, ArgsTable } from "@storybook/addon-docs/blocks";
 
 import Tooltip from "../labsystem/src/Tooltip";
-import OutlineButton from "../labsystem/src/Buttons/OutlineButton";
+import OutlineButton from "../labsystem/src/Button/OutlineButton";
 import Icon from "../labsystem/src/Icon";
 
 <Meta title="Components/Tooltip" />
